# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: >
    API for the WASAText messaging application, which allows users to send and receive
    individual or group messages with support for text, photos, and reactions.
    Includes simplified login, user management, conversations, messages, and group handling.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "user-id"
  schemas:
    ErrorResponse:
      type: object
      description: Standard error response payload.
      properties:
        error:
          type: string
          description: Human-readable error message describing what went wrong.
          example: "An error occurred."
          minLength: 1
          maxLength: 255
          pattern: '^.+$'
      required:
        - error
    User:
      type: object
      description: Represents an application user.
      properties:
        id:
          type: string
          description: Unique user identifier.
          example: "user123"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        name:
          type: string
          description: Display name of the user.
          example: "Maria"
          minLength: 1
          maxLength: 50
          pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,50}$'
        photo:
          type: string
          format: uri
          description: Public URL of the user profile photo.
          example: "https://example.com/maria.jpg"
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+$'
      required: [name]
    Conversation:
      type: object
      description: Represents a conversation (direct or group).
      properties:
        id:
          type: string
          description: Unique conversation identifier.
          example: "conv123"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        type:
          type: string
          description: Type of conversation (direct or group).
          enum: [direct, group]
          example: "group"
          minLength: 5
          maxLength: 6
          pattern: '^(direct|group)$'
        name:
          type: string
          description: Human-readable name of the conversation (mainly for groups).
          example: "Work Friends"
          minLength: 1
          maxLength: 100
          pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,100}$'
        photo:
          type: string
          format: uri
          description: Public URL of the conversation or group photo.
          example: "https://example.com/group-photo.jpg"
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+$'
        lastMessagePreview:
          type: string
          description: Short preview of the most recent message in the conversation.
          example: "See you tomorrow?"
          minLength: 0
          maxLength: 140
          pattern: '^[\s\S]*$'
        lastMessageAt:
          type: string
          format: date-time
          description: Timestamp of the most recent message in the conversation (ISO 8601).
          example: "2025-11-04T20:00:00Z"
        lastMessageSenderId:
          type: string
          description: ID of the user who sent the last message.
          example: "user123"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        lastMessageSenderName:
          type: string
          description: Name of the user who sent the last message.
          example: "Juan"
          minLength: 1
          maxLength: 50
          pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,50}$'
        lastMessageStatus:
          type: string
          description: Status of the last message.
          enum: [delivered, read]
          example: "read"
          minLength: 4
          maxLength: 9
          pattern: '^(delivered|read)$'
        unreadCount:
          type: integer
          description: Number of unread messages for the requesting user.
          example: 5
          minimum: 0
        participants:
          type: array
          description: List of users participating in the conversation.
          minItems: 1
          maxItems: 512
          items:
            $ref: "#/components/schemas/User"
      required: [id, type, participants]
    Reaction:
      type: object
      description: Represents a reaction from a user to a message.
      properties:
        id:
          type: string
          description: Unique reaction identifier.
          example: "react789"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        user:
          $ref: "#/components/schemas/User"
        emoji:
          type: string
          description: Emoji used for the reaction.
          example: "😂"
          minLength: 1
          maxLength: 8
          pattern: '^.+$'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the reaction was added (ISO 8601).
          example: "2025-11-08T15:00:00Z"
      required: [id, user, emoji, createdAt]
    MessageSnapshot:
      type: object
      description: A simplified view of a message used in replies/quotes.
      properties:
        id:
          type: string
          description: The ID of the original message (might not exist anymore).
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        text:
          type: string
          description: Text content snapshot.
          minLength: 0
          maxLength: 4096
          pattern: '^[\s\S]*$'
        photo:
          type: string
          format: uri
          description: Photo URL snapshot.
          minLength: 0
          maxLength: 2048
          pattern: '^$|https?://.+$'
        sender:
          type: object
          description: Only contains the sender's name at the time of the snapshot.
          properties:
            name:
              type: string
              example: "Maria"
              minLength: 1
              maxLength: 50
              pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,50}$'
    Message:
      type: object
      description: Represents a message sent in a conversation.
      properties:
        id:
          type: string
          description: Unique message identifier.
          example: "msg001"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        sender:
          $ref: "#/components/schemas/User"
        conversationId:
          type: string
          description: Identifier of the conversation where this message belongs.
          example: "conv123"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        text:
          type: string
          description: Text content of the message, if any.
          example: "Hello 👋"
          minLength: 0
          maxLength: 4096
          pattern: '^[\s\S]*$'
        photo:
          type: string
          format: uri
          description: Optional URL of a photo attached to the message.
          example: "https://example.com/image.jpg"
          minLength: 0
          maxLength: 2048
          pattern: '^$|https?://.+$'
        replyToMessageId:
          type: string
          description: Identifier of the message being replied to, if applicable.
          example: "msg000"
          minLength: 0
          maxLength: 64
          pattern: '^$|[A-Za-z0-9_-]+$'
        replyTo:
          $ref: "#/components/schemas/MessageSnapshot"
          
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the message was created (ISO 8601).
          example: "2025-11-04T10:30:00Z"
        reactions:
          type: array
          description: List of reactions users have added to this message.
          minItems: 0
          maxItems: 100
          items:
            $ref: "#/components/schemas/Reaction"
        status:
          type: string
          description: >
            Message status:
            - delivered: shown as a single checkmark (delivered to all recipients)
            - read: shown as a double checkmark (read by all recipients)
          enum: [delivered, read]
          example: "read"
          minLength: 4
          maxLength: 9
          pattern: '^(delivered|read)$'
      required: [id, sender, conversationId, createdAt, status]
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      description: Conversation ID.
      schema:
        type: string
        example: "conv123"
        minLength: 1
        maxLength: 64
        pattern: '^[A-Za-z0-9_-]+$'
    MessageId:
      name: messageId
      in: path
      required: true
      description: Message ID.
      schema:
        type: string
        example: "msg001"
        minLength: 1
        maxLength: 64
        pattern: '^[A-Za-z0-9_-]+$'
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List or search users
      description: >
        Retrieve a list of users. Can filter by name (prefix search).
        The authenticated user is excluded from the results.
      operationId: listOrSearchUsers
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          description: Optional string to filter users (names starting with this value).
          required: false
          schema:
            type: string
            example: "Mar"
            minLength: 0
            maxLength: 32
      responses:
        '200':
          description: List of users found.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /conversations:
    get:
      tags:
        - Conversations
      summary: List user conversations
      description: >
        Retrieve all conversations the user participates in.
        Results are sorted by most recent activity (last message timestamp or join date).
        For 'direct' conversations, the name and photo returned are those of the other participant.
      operationId: getMyConversations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of conversations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '500':
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Groups
      summary: Create a new group
      description: Create a new group conversation. The authenticated user is automatically added as the creator.
      operationId: createGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload for creating a group.
              properties:
                name:
                  type: string
                  description: Name of the new group.
                  example: "Work Friends"
                  minLength: 1
                  maxLength: 100
                  pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,100}$'
                users:
                  type: array
                  description: List of user IDs to invite.
                  minItems: 0
                  maxItems: 511
                  items:
                    type: string
                    description: User identifier.
                    minLength: 1
                    maxLength: 64
                    pattern: '^[A-Za-z0-9_-]+$'
                  example: ["user123", "user456"]
              required:
                - name
                - users
      responses:
        '201':
          description: Group created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /conversations/{conversationId}/messages/{messageId}:
    parameters:
      - $ref: "#/components/parameters/ConversationId"
      - $ref: "#/components/parameters/MessageId"

    get:
      summary: Get a specific message
      description: Retrieve a single message from a conversation by its identifier.
      operationId: getMessageById
      tags:
        - Messages
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Message retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          description: Missing conversation or message ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          description: User is not a participant of the conversation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Conversation or message not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Messages
      summary: Forward a message
      description: >
        Forward an existing message to another conversation.
        A new message is created in the target conversation with the same content (text/photo).
        The user must be a participant of both the source and target conversations.
      operationId: forwardMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload indicating the target conversation.
              properties:
                targetConversationId:
                  type: string
                  description: ID of the conversation to forward the message to.
                  example: "conv456"
                  minLength: 1
                  maxLength: 64
              required:
                - targetConversationId
      responses:
        '201':
          description: Message forwarded successfully (returns the new message).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          description: Invalid JSON or missing targetConversationId.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          description: User is not a participant of the source or target conversation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Source conversation or message not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Messages
      summary: Delete a message
      description: >
        Delete a message previously sent.
        **Restriction:** A user can only delete messages they sent themselves.
      operationId: deleteMessage
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Message deleted successfully.
        '400':
          description: Missing conversation or message ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          description: User is not the sender of this message or not in the conversation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Conversation or message not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /conversations/{conversationId}/messages/{messageId}/reactions:
    parameters:
      - $ref: "#/components/parameters/ConversationId"
      - $ref: "#/components/parameters/MessageId"

    get:
      tags:
        - Messages
      summary: List reactions for a message
      description: Retrieve all reactions added to a specific message.
      operationId: listReactionsForMessage
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Reactions retrieved successfully. Returns an empty list if the message has no reactions or does not exist.
          content:
            application/json:
              schema:
                type: array
                description: Array of reactions.
                items:
                  $ref: "#/components/schemas/Reaction"
        '400':
          description: Missing conversation or message ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          description: User is not in the conversation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Messages
      summary: Add or update a reaction
      description: >
        Add a reaction to a specific message.
        **Note:** If the user has already reacted to this message, the existing reaction
        will be updated with the new emoji (a user can only have one reaction per message).
      operationId: commentMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload specifying the emoji.
              properties:
                emoji:
                  type: string
                  description: Emoji to attach.
                  example: "👍"
                  minLength: 1
                  maxLength: 8
              required:
                - emoji
      responses:
        '201':
          description: Reaction added or updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
        '400':
          description: Invalid JSON or missing emoji.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          description: User is not in the conversation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Conversation or message not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          $ref: "#/components/responses/InternalServerError"
...